Attempt to combine directive specs for cost and listSize into a single definition
