+++ composition-js/src/composeDirectiveManager.ts
@@ -67,7 +67,6 @@ const DISALLOWED_IDENTITIES = [
   'https://specs.apollo.dev/source',
   'https://specs.apollo.dev/context',
   'https://specs.apollo.dev/cost',
-  'https://specs.apollo.dev/listSize',
 ];
 
 export class ComposeDirectiveManager {
