+++ internals-js/src/index.ts
@@ -26,4 +26,3 @@ export * from './specs/requiresScopesSpec';
 export * from './specs/policySpec';
 export * from './specs/sourceSpec';
 export * from './specs/costSpec';
-export * from './specs/listSizeSpec';
