+++ internals-js/src/specs/federationSpec.ts
@@ -21,7 +21,6 @@ import { POLICY_VERSIONS } from './policySpec';
 import { SOURCE_VERSIONS } from './sourceSpec';
 import { CONTEXT_VERSIONS } from './contextSpec';
 import { COST_VERSIONS } from "./costSpec";
-import { LIST_SIZE_VERSIONS } from "./listSizeSpec";
 
 export const federationIdentity = 'https://specs.apollo.dev/federation';
 
@@ -189,7 +188,6 @@ export class FederationSpecDefinition extends FeatureDefinition {
 
     if (version.gte(new FeatureVersion(2, 9))) {
       this.registerSubFeature(COST_VERSIONS.find(new FeatureVersion(0, 1))!);
-      this.registerSubFeature(LIST_SIZE_VERSIONS.find(new FeatureVersion(0, 1))!);
     }
   }
 }
